+++ lib/rubocop/cop/style/for.rb
@@ -4,7 +4,7 @@ module Rubocop
   module Cop
     module Style
       # This cop looks for uses of the *for* keyword.
-      class AvoidFor < Cop
+      class For < Cop
         MSG = 'Prefer *each* over *for*.'
 
         def on_for(node)