+++ spec/rubocop/cop/style/for_spec.rb
@@ -5,31 +5,27 @@ require 'spec_helper'
 module Rubocop
   module Cop
     module Style
-      describe AvoidFor do
-        subject(:af) { AvoidFor.new }
+      describe For do
+        subject(:cop) { described_class.new }
 
         it 'registers an offence for for' do
-          inspect_source(af,
+          inspect_source(cop,
                          ['def func',
                           '  for n in [1, 2, 3] do',
                           '    puts n',
                           '  end',
                           'end'])
-          expect(af.offences.size).to eq(1)
-          expect(af.messages)
-            .to eq([AvoidFor::MSG])
+          expect(cop.offences.size).to eq(1)
         end
 
         it 'does not register an offence for :for' do
-          inspect_source(af,
-                         ['[:for, :ala, :bala]'])
-          expect(af.offences).to be_empty
+          inspect_source(cop, ['[:for, :ala, :bala]'])
+          expect(cop.offences).to be_empty
         end
 
         it 'does not register an offence for def for' do
-          inspect_source(af,
-                         ['def for; end'])
-          expect(af.offences).to be_empty
+          inspect_source(cop, ['def for; end'])
+          expect(cop.offences).to be_empty
         end
       end
     end